+++ src/assets/jss/material-dashboard-react/components/customInputStyle.jsx
@@ -18,7 +18,8 @@ const customInputStyle = {
     },
     "&:after": {
       borderColor: primaryColor
-    }
+    },
+    color: "#fff"
   },
   underlineError: {
     "&:after": {