+++ src/assets/jss/material-dashboard-react/components/sidebarStyle.jsx
@@ -178,6 +178,16 @@ const sidebarStyle = theme => ({
         "0 12px 20px -10px rgba(0,188,212,.28), 0 4px 20px 0 rgba(0,0,0,.12), 0 7px 8px -5px rgba(0,188,212,.2)"
     }
   },
+  black: {
+    backgroundColor: primaryColor,
+    boxShadow:
+      "0 12px 20px -10px rgba(100, 100, 100, 1), 0 4px 20px 0 rgba(0,0,0,0), 0 7px 8px -5px rgba(100, 100, 100, .12)",
+    "&:hover": {
+      backgroundColor: primaryColor,
+      boxShadow:
+        "0 12px 20px -10px rgba(100, 100, 100, 1), 0 4px 20px 0 rgba(0,0,0,0), 0 7px 8px -5px rgba(100, 100, 100, .12)"
+    }
+  },
   green: {
     backgroundColor: successColor,
     boxShadow: