+++ src/layouts/Dashboard/Dashboard.jsx
@@ -76,7 +76,7 @@ class App extends React.Component {
           image={image}
           handleDrawerToggle={this.handleDrawerToggle}
           open={this.state.mobileOpen}
-          color="blue"
+          color="black"
           {...rest}
         />
         <div className={classes.mainPanel} ref="mainPanel">